feat: remove "negative tolerance" feature
We found that this part of the algorithm often lead to the result that groups ordering a full bunch of a product had to give 1 or 2 units to other groups.
As we considered this an unwanted effect rendering the product distribution more laborious and complicated - without providing any benefit that we'd be aware of (maybe we just couldn't get the original authors wise idea), we decided to take out the feature entirely.